chore: restructure whatsapp code and fix send template task
